DigitalAI/Social_Media/requirements.txt
flask==3.0.2
streamlit==1.31.1
openai==1.12.0
python-dotenv==1.0.1
flask-cors==4.0.0
requests==2.31.0